+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes from ns-3.41 to ns-3-dev
 --------------------------------
 
 ### New API
+* (wifi) New trace sources `SpectrumWifiPhy::SignalTransmission`, `YansWifiPhy::SignalArrival`, and `YansWifiPhy::SignalTransmission`
 
 * Objects now can be aggregated to multiple objects though the `Object::UnidirectionalAggregateObject` function. Objects aggregated in such a way can not use `GetObject` to access the objects they are aggregated to.
 * (network) Added `ApplicationHelper` helper class to create and install applications, removing